Initial player service interop

+++ shared/src/commonMain/kotlin/de/julianostarek/motif/feed/datasource/FeedLocalDataSource.kt
@@ -5,5 +5,5 @@ import de.julianostarek.motif.feed.dto.FeedMotifDto
 interface FeedLocalDataSource : FeedDataSource {
     suspend fun dumpMotifsFeed(motifsFeed: List<FeedMotifDto>)
     suspend fun insertMotif(motif: FeedMotifDto)
-    suspend fun deleteMotif(motifId: Long)
+    suspend fun deleteMotif(motifId: Int)
 }
